- fix issue casting tinyint from sp - invoke right method to create menuupdated payload

+++ StarChef.Orchestrate/Models/Menu.cs
@@ -32,7 +32,7 @@
                 .SetCustomerName(cust.Name)
                 .SetExternalId(reader[1].ToString())
                 .SetMenuName(reader[2].ToString())
-                .SetMenuType((int)reader[3] == 1 ? Events.MenuType.ALACARTE : Events.MenuType.BUFFET)
+                .SetMenuType((byte)reader[3] == 1 ? Events.MenuType.ALACARTE : Events.MenuType.BUFFET)
                 .SetSource(Events.SourceSystem.STARCHEF)
                 .SetSequenceNumber(rand.Next(1, int.MaxValue));
             }
